# ARROW-14577: [C++] Enable fine grained IO for async IPC reader 