In the end this PR only addresses projection pushdown on the asynchronous path.  The approach laid out here could be used on the synchronous path to offer pre-buffering.  Tests should be done to see if that is faster.

The memory mapped / will need issue should be solved by making auto-will-need a property of the memory mapped file and will not be addressed here.