@westonpace 2021-12-04
This is ready for review but at the moment the performance is worse in the micro-benchmark.  The micro-benchmark is dealing with a 1MB in-memory file so my guess is the fact that we have to make more read calls for partial reads is hurting.  I will try it again with larger files with memory contention and with disk I/O to see if that changes anything.