@westonpace 2021-12-08
At a macro level at least it seems there is some benefit, but definitely some strange behaviors going on:

![async](https://user-images.githubusercontent.com/1696093/145145405-3e1b6ebf-35e5-456d-b299-4625e82ed775.png)
![old](https://user-images.githubusercontent.com/1696093/145145409-859f8a57-36d7-4f21-9765-8cd50e4bf67d.png)
![sync](https://user-images.githubusercontent.com/1696093/145145411-297350d5-d11a-4a1c-8b54-c7d1751912a8.png)