@westonpace 2021-12-08
(fixed charts).  There is some benefit at the macro-level.  These are hot I/O reads which explains why we are able to see some parallelism:

![async](https://user-images.githubusercontent.com/1696093/145145590-1167fc33-6d81-4545-b70f-03d5a539db0a.png)
![old](https://user-images.githubusercontent.com/1696093/145145591-7da6c6be-33f6-4583-901d-a056b4cd181e.png)
![sync](https://user-images.githubusercontent.com/1696093/145145593-11be46a0-060a-4918-b9da-724fee6ffd14.png)