@westonpace 2021-12-08
Similar results with cold I/O
![async-cold](https://user-images.githubusercontent.com/1696093/145146046-2dd47ccb-60d0-4bad-aa6b-60a260d7f324.png)
![old-cold](https://user-images.githubusercontent.com/1696093/145146049-dad8da50-2256-4847-a963-3709bb8f5e76.png)
![sync-cold](https://user-images.githubusercontent.com/1696093/145146052-5af140ae-b919-4a24-92e0-ad0105d270ec.png)